requirements.txt
ELM327-emulator
Flask
requests
